Adding a PATH definition to the crontab entry
On upgrade to EngineYard stable-v4 stack and Ruby v2 the running of
bundle from the cron job was failing (#4046).

This was due to the fact the bundle was no longer installed under
/usr/bin.  See the EY ticket here

https://support.cloud.engineyard.com/requests/68021

The solution was to redefine PATH as part of the cron task so that it
includes /usr/local/bin, where bundler can be found.