Updating Sidekiq Connection Pool
Updating the sidekiq connection pool to take into account concurrency and number of workers when setting the connection pool.

Previously, the connection pool was set to the concurrency without any regard for the number of works. So, 8 works with a concurrecy of 25 woudl be 200 Sidekiq threads, but it would only have a connection pool of 25.